+++ eva/src/components/dashBoard.vue
@@ -296,6 +296,24 @@
                     <div class="settings-dash-block">
                       <div class="settings-dash">
                         <v-tooltip
+                          v-if="!excludedFromDataSearches && elementType === 'multiLine'"
+                          bottom
+                          :color="theme.$accent_ui_color"
+                          :open-delay="tooltipOpenDelay"
+                        >
+                          <template v-slot:activator="{ on }">
+                            <v-icon
+                              class="icon"
+                              :color="theme.$main_border"
+                              v-on="on"
+                              @click="screenShot($refs.fullScreenCard.$el)"
+                            >
+                              {{ props.mdiCamera }}
+                            </v-icon>
+                          </template>
+                          <span>Скачать в PNG</span>
+                        </v-tooltip>
+                        <v-tooltip
                           v-if="isMultiline"
                           bottom
                           :color="theme.$accent_ui_color"
@@ -360,7 +378,11 @@
                   </v-card-title>
 
                   <portal-target
+                    ref="fullScreenCard"
                     style="height: calc(100% - 35px)"
+                    :style="{
+                      background: theme.$main_bg
+                    }"
                     :name="`${element}`"
                   />
                 </v-card>
@@ -372,6 +394,24 @@
             class="settings-dash"
             :class="{ settings_move: props.open_gear }"
           >
+            <v-tooltip
+              v-if="!excludedFromDataSearches && elementType === 'multiLine'"
+              bottom
+              :color="theme.$accent_ui_color"
+              :open-delay="tooltipOpenDelay"
+            >
+              <template v-slot:activator="{ on }">
+                <v-icon
+                  class="icon"
+                  :color="theme.$main_border"
+                  v-on="on"
+                  @click="screenShot($refs.screenCard.$el)"
+                >
+                  {{ props.mdiCamera }}
+                </v-icon>
+              </template>
+              <span>Скачать в PNG</span>
+            </v-tooltip>
             <v-tooltip
               v-if="!excludedFromDataSearches"
               bottom
@@ -486,7 +526,7 @@
         </div>
       </div>
       <v-card-text
-        v-if="!excludedFromDataSearches && currentElem !== 'dash-constructorSchemes'"
+        v-if="!excludedFromDataSearches"
         v-show="!showElement"
         class="card-text"
       >
@@ -500,13 +540,14 @@
       </v-card-text>
       <v-card-text
         :is="currentElem"
-        v-if="showElement || excludedFromDataSearches || currentElem === 'dash-constructorSchemes'"
+        v-if="showElement || excludedFromDataSearches"
+        ref="screenCard"
         :full-screen-mode="bigSizeMode"
         custom-class="card-text element-itself"
         :color-from="theme"
         :custom-style="{
           color: theme.$main_text,
-          background: 'transparent'
+          background: theme.$main_bg
         }"
         :id-from="element"
         :id-dash-from="idDash"
@@ -566,6 +607,7 @@ import {
   mdiTrashCanOutline,
   mdiFullscreen,
   mdiFullscreenExit,
+  mdiCamera,
 } from '@mdi/js';
 import settings from '../js/componentsSettings';
 import visualisation from '../js/visualisationCRUD';
@@ -651,6 +693,7 @@ export default {
         mdiArrowCollapse,
         mdiFullscreen,
         mdiFullscreenExit,
+        mdiCamera,
         icons: {},
         edit: true,
         edit_icon: true,
@@ -1341,6 +1384,22 @@ export default {
     toggleBigSize() {
       this.bigSizeMode = !this.bigSizeMode;
       this.isFullScreen = false;
+    },
+    screenShot(screen) {
+      this.$html2canvas(screen, { type: 'dataURL' }).then((canvas) => {
+        this.downloadTheme(canvas);
+      });
+    },
+    downloadTheme(canvas, postfix = (+(new Date())).toString()) {
+      const lnk = document.createElement('a');
+      lnk.href = canvas;
+      lnk.download = this.element;
+      lnk.target = '_blank';
+      lnk.style.display = 'none';
+      lnk.id = `downloadlnk-${postfix}`;
+      document.body.appendChild(lnk);
+      lnk.click();
+      document.body.removeChild(lnk);
     },
   },
 };