+++ eva/src/js/componentsSettings.js
@@ -42,13 +42,14 @@ export default {
     { name: 'Текстовый блок', img: mdiCardTextOutline, type: 'textarea' },
     { name: 'Граф_old', img: mdiGraph, type: 'graph' },
     { name: 'Тепловая карта', img: mdiGrid, type: 'heatmap' },
-    { name: 'Single Value', img: mdiNumeric, type: 'singleValue' },
+    { name: 'Показатели', img: mdiNumeric, type: 'singleValue' },
     { name: 'Ползунок', img: 'eva-basic_slider_01', type: 'tune' },
     { name: 'Конструктор схем', img: mdiTuneVertical, type: 'constructorSchemes' },
     { name: 'Накопитель', img: 'eva-chart_bar_chart_horizontal', type: 'accumulators' },
     { name: 'Меню', img: 'eva-edit_list_checklist', type: 'menu' },
     { name: 'Круговая шкала', img: mdiImageFilterTiltShift, type: 'dial' },
     { name: 'Точечный график', img: mdiScatterPlotOutline, type: 'scatterPlot' },
+    { name: 'Динамическая форма', img: mdiScatterPlotOutline, type: 'dynamicForm' },
   ],
   size: {
     picker: {
@@ -143,6 +144,10 @@ export default {
       width: 930,
       height: 850,
     },
+    dynamicForm: {
+      width: 400,
+      height: 400,
+    },
   },
   icons: {
     table: mdiTableLarge,
@@ -168,6 +173,7 @@ export default {
     menu: 'eva-edit_list_checklist',
     dial: mdiImageFilterTiltShift,
     scatterPlot: mdiScatterPlotOutline,
+    dynamicForm: mdiScatterPlotOutline,
   },
   commonOptions: [
     'panelSettings',
@@ -212,7 +218,13 @@ export default {
       'defaultSourceDataField',
       'defaultSourceDataUpdates',
     ],
-    picker: [],
+    picker: [
+      'showLastTimeBlock',
+      'showChoseDateAndTimeBlock',
+      'showRangeDateBlock',
+      'showCustomInputBlock',
+      'timeOutputFormat',
+    ],
     graph: ['boxShadow'],
     single: [
       'subnumber',
@@ -242,6 +254,7 @@ export default {
       'validationType',
       'validationNumberRangeMin',
       'validationNumberRangeMax',
+      'readOnly',
     ],
     guntt: ['timeFormat'],
     tile: [
@@ -255,7 +268,7 @@ export default {
     csvg: ['tooltip', 'backgroundImage'],
     ygraph: [],
     bush: [],
-    map: ['osmserver', 'primitivesLibrary'],
+    map: ['osmserver', 'primitives', 'primitivesLibrary'],
     heatmap: [
       'dataFormat',
       'x',
@@ -294,7 +307,8 @@ export default {
       'metricGroup',
       'scatterPlotLegend',
     ],
-    constructorSchemes: ['visible', 'level', 'pinned'],
+    constructorSchemes: ['visible', 'level', 'pinned', 'primitives', 'primitivesLibrary'],
+    dynamicForm: ['visible', 'level', 'pinned', 'formGenerator', 'formOptions'],
   },
   optionFields: [
     // описание типов полей и их характеристик
@@ -395,6 +409,12 @@ export default {
       elem: 'switch',
     },
     {
+      option: 'readOnly',
+      description: 'Только чтение',
+      elem: 'switch',
+      default: false,
+    },
+    {
       option: 'textFontSize',
       description: 'Выбрать размер шрифта',
       elem: 'select',
@@ -485,6 +505,39 @@ export default {
       description:
         'Сервер для набора tile Пример:\nhttps://tile.openstreetmap.org/{z}/{x}/{y}.png',
       elem: 'text-field',
+    },
+
+    // datepicker
+    {
+      option: 'showLastTimeBlock',
+      description: 'Показать блок: Выбор времени',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'showChoseDateAndTimeBlock',
+      description: 'Показать блок: Выбор времени и даты',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'showRangeDateBlock',
+      description: 'Показать блок: Диапазон дат',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'showCustomInputBlock',
+      description: 'Показать блок: Ввод даты и времени вручную',
+      elem: 'switch',
+      default: true,
+    },
+    {
+      option: 'timeOutputFormat',
+      description: 'Формат даты для результата',
+      elem: 'text-field',
+      default: '',
+      placeholder: 'Пример: YYYY-MM-DD HH:mm',
     },
 
     // dashSingle
@@ -629,7 +682,6 @@ export default {
         return this.$store.state[this.idDash][this.element]?.availableTableTitles;
       },
     },
-
     {
       option: 'columnCount',
       description: 'Количество столбцов',
@@ -788,6 +840,30 @@ export default {
       elem: 'switch',
     },
 
+    // fullWidthGroup
+    {
+      group: 'Библиотека примитивов',
+      option: 'primitives',
+    },
+    {
+      optionGroup: 'primitives',
+      option: 'primitivesLibrary',
+      elem: 'code-editor',
+      isFullWidth: true,
+    },
+    // formGenerator
+    {
+      group: 'Исходные данные для формы',
+      option: 'formGenerator',
+    },
+    {
+      optionGroup: 'formGenerator',
+      option: 'formOptions',
+      elem: 'code-editor',
+      isFullWidth: true,
+      importantExport: true,
+    },
+
   ],
   reporstElements: [
     'table',
@@ -850,7 +926,7 @@ export default {
       icon: mdiGrid,
     },
     singleValue: {
-      tooltip: 'Single Value',
+      tooltip: 'Показатели',
       icon: mdiNumeric,
     },
     tune: {
@@ -869,6 +945,8 @@ export default {
     fromDataSearches: [
       'menu',
       'picker',
+      'constructorSchemes',
+      'dynamicForm',
     ],
   },
 };