+++ RELEASENOTES.md
@@ -1,3 +1,25 @@
+# [1.108.0] - 2022-10-13
+
+### Added
+- Added new dash Dynamic forms
+- Data Picker: Added settings for displaying blocks of the entered date and time
+- Data Picker: Added the ability to format the result
+- Single Value: Added the ability to assign an icon via metadata
+- Single Value: Added ability to correct color via metadata
+- Constructor schemes: Added code editor for primitives library field
+- Constructor schemes: Added loading svg icons from server
+- Constructor schemes: Added data source support
+- Added possibility save multiLine in png
+- Text Area: added read only mode
+
+### Changed
+- Renaming the Single Value component to 'Показатели'
+
+### Fixed
+- fixed update of tokens on the map
+- fixed position icons\buttons on dash panel
+- fixed flickering map
+
 # [1.107.0] - 2022-09-08
 
 ### Added